feat: enhance error handling and logging in various components, improve parameter naming for clarity, and add port availability check in test scripts

+++ proj2/src/polynomial.h
@@ -48,7 +48,7 @@ class Polynomial {
     double score() const {
         double score = 0.0;
         for (size_t i = 0; i < points.size(); ++i) {
-            score += squared_error(points[i], evaluate(i));
+            score += squared_error(points[i], evaluate((double)i));
         }
 
         return score;
